added comments updated functions name

+++ backend/src/plugin/api/api.controller.ts
@@ -4,20 +4,30 @@ import { Collection } from "mongodb";
 import { Collections } from "../../collections";
 import { Phrase } from "../../models/file.model";
 import fs from "fs/promises";
+import { ResponseObject } from "hapi";
 export default class ApiController {
   constructor() {}
 
-  public async writePhraseToMongo(request: IExtendedRequest) {
+  // function to write phrase into Mongo
+  public async writePhraseToMongo(
+    request: IExtendedRequest,
+    h: ResponseObject
+  ) {
     try {
       // grab db from app namespace
 
-      console.log("in write to mongo", request.payload);
       const db = request.server.app.db;
       const collection: Collection = db.collection(Collections.Phrase);
       const payload = request.payload as any;
       const id = payload.id;
       const phrase = payload.phrase;
-      console.log("id phrase is ", id, phrase);
+
+      // if id or phrase is not present
+      if (!id || !phrase) {
+        return h.code(400);
+      }
+
+      // create phrase
       const newPhrase: Phrase = { id, phrase };
       const result = await collection.insertOne(newPhrase);
 
@@ -30,12 +40,20 @@ export default class ApiController {
     }
   }
 
-  public async deletePhareFromMongo(request: IExtendedRequest) {
+  // function to delete phrase from mongo
+  public async deletePhareFromMongo(
+    request: IExtendedRequest,
+    h: ResponseObject
+  ) {
     try {
       const db = request.server.app.db;
       const id = request.params.id as any;
-      console.log("id is", id);
       const collection: Collection = db.collection(Collections.Phrase);
+      // checking if phrase with id exists or not
+      const phrase = await collection.findOne({ id: Number(id) });
+      if (!phrase) return h.code(400);
+
+      // if exist delete it
       const result = await collection.deleteOne({
         id: Number(id),
       });
@@ -49,6 +67,7 @@ export default class ApiController {
     }
   }
 
+  // function to load file data to mongo
   public async loadPhareToMongo(request: IExtendedRequest) {
     try {
       let data = await fs.readFile(__dirname + "/file.txt");
@@ -64,6 +83,8 @@ export default class ApiController {
 
       const db = request.server.app.db;
       const collection: Collection = db.collection(Collections.Phrase);
+
+      // performing bulk write
       const result = await collection.insertMany(res);
 
       return {
@@ -92,7 +113,9 @@ export default class ApiController {
       return Boom.boomify(err);
     }
   }
-  public async getPhrase(request: IExtendedRequest) {
+
+  // function to get phrases directly from file
+  public async getPhraseFromFile(request: IExtendedRequest) {
     try {
       let data = await fs.readFile(__dirname + "/file.txt");
       let sData = data.toString().split("\n");
@@ -114,13 +137,19 @@ export default class ApiController {
     }
   }
 
-  public async write(request: IExtendedRequest) {
+  // function to write phrase in file
+  public async writePhraseInFile(request: IExtendedRequest) {
     try {
+      // read all phrases
       let data = await fs.readFile(__dirname + "/file.txt");
       let sData = data.toString().split("\n");
+
+      // computing write line number
       let lineNumber = sData.length;
       const payload = request.payload as any;
       const phrase = payload.phrase;
+
+      // writing phrase in file
       await fs.appendFile(`${__dirname}/file.txt`, `${phrase} \n`);
       let res = { id: lineNumber + 1, phrase };
       return {
@@ -132,25 +161,34 @@ export default class ApiController {
       return Boom.boomify(err);
     }
   }
-  public async delete(request: IExtendedRequest) {
+  public async deleteFromFile(request: IExtendedRequest) {
     try {
       const payload = request.payload as any;
       const id = request.params.id as any;
+
+      // rading from file
       let data = await fs.readFile(__dirname + "/file.txt");
       let sData = data.toString().split("\n");
+
+      // checking if delete line number is valid or not
       if (id > sData.length - 1) {
         return {
           message: "Delete line Not Possible",
           data: { success: false },
         };
       }
+
+      // removing line number
       sData.splice(id - 1, 1);
       console.log("data is ", sData);
       let writeData = "";
+
+      // aggregrating all phrases
       sData.map((data) => {
         writeData += data + "\n";
       });
 
+      // writing content in file
       await fs.writeFile(__dirname + "/file.txt", writeData);
       return {
         message: "Delete phrase  success",
